curate_bot/requirements.txt
selenium==4.11.2
pillow==10.0.0
opencv-python==4.8.0.74
pytesseract==0.3.10
notion-client==2.0.0
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.3.0
numpy==1.24.3
datetime==5.1
argparse==1.4.0
watchdog
google-api-python-client

# pip install -r requirements.txtで実行
